pyproject.toml
[tool.poetry]
name = "yellow-pokemon"
version = "0.1.0"
description = "Instagram Clone"
authors = ["yellow"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
Pillow = "^8.3.1"
django-environ = "^0.4.5"
django-simple-chatbot = "^0.0.9"
djangorestframework = "^3.12.4"
nltk = "^3.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
